requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
colorama==0.4.4
discord-py-interactions==4.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
orjson==3.6.7
py-cord==2.0.0b5
yarl==1.7.2
